requirements.txt
Flask
pyTelegramBotAPI
groq
schedule
pymongo
pytz
python-dotenv